requirements.txt
discord == 2.2.2
python-dotenv == 1.0.0
mysql-connector-python == 8.0.32